mdns: Add MDNS lookup ability to base Supervisor image
This includes `libnss-mdns` and implicitly Avahi for MDNS
lookups. It also patches the Avahi recipe to forgo the use
of `gobject-introspection-data` which has problems and is
now no longer recommended for native builds. This also
removes the feature from the distro.

Also adds a brief README, for Concourse compliance.

Connects-to: #9
Change-type: minor
Signed-off-by: Heds Simons <heds@resin.io>