meta-balena-supervisor/nodejs: Restructure and fix version 6
First of all we fix the path of the inc file required by
nodejs_6.13.1.bb by specifying a correct FILESEXTRAPATHS relative to the
recipe's directory.
We also add the DNS patch only for the 6 version as it is included in
version 8 (which is the other version we are using).
The last thing included in this patch is fixing compilation by using
openssl 1.0.x for nodejs 6 because this version doesn't support openssl
1.1.x. See: https://github.com/nodejs/node/pull/16130

Signed-off-by: Andrei Gherzan <andrei@balena.io>